Add workers on loading
