+++ src/components/index.scss
@@ -159,6 +159,17 @@ body {
     letter-spacing: 1px;
     user-select: none;
 
+    &::first-letter {
+        text-transform: uppercase;
+    }
+
+    blockquote {
+        &:after {
+            content: '.';
+            display: inline;
+        }
+    }
+
     &:before{
         position: absolute;
         content: '\201C';
@@ -284,7 +295,6 @@ body {
 .box{
     width: 50%;
     height: 60%;
-    border: 1px solid $orange-colour;
 }
 
 .loading-title{
@@ -294,11 +304,59 @@ body {
     text-align: center;
     user-select: none;
 }
-.loading-text{
+.loading-image{
+    width: 30%;
+    margin: 7% auto 0 auto;
+    fill:#ffffff;
+}
+.loading-text, .loading-about {
     color: white;
-    font-size: 1.5em;
+    font-size: 1.2em;
     text-align: center;
     user-select: none;
+    padding: .5em 0 1.5em 0;
+}
+
+.loading-about {
+    font-size: 1.5em;
+    padding-top: 1.5em;
+    border-top: solid 1px #ffffff77;
+}
+
+.galaxy {
+    @keyframes spin {
+        0% {
+            transform: rotate(0deg); 
+        }
+
+        100% { 
+            transform: rotate(360deg); 
+        }
+    }
+
+    & > * {
+        transform-origin: 256px 256px;
+    }
+
+    $base-speed: 10s;
+
+    .sun {
+        fill:$orange-colour;
+        animation: spin $base-speed linear infinite;
+    }
+
+    .inner-arc {
+        animation: spin ($base-speed * 3) linear infinite reverse;
+    }
+
+    .outter-arc {
+        animation: spin ($base-speed * 4) linear infinite;
+    }
+
+    .stars {
+        fill: $orange-colour;
+        animation: spin ($base-speed * 4) linear infinite;
+    }
 }
 
 .arrow {
