+++ src/index.jsx
@@ -1,15 +1,27 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Layout from './components/Layout';
-import Loading from './Loading';
+import Loading from './components/Loading';
+import PredictionWorker from 'workerize-loader!./workers/predictor'; // eslint-disable-line import/no-webpack-loader-syntax
 
 class App extends React.Component {
+    constructor() {
+        super();
+
+        this.state = {messages: null};
+
+        this.predictor = new PredictionWorker();
+
+        this.predictor.predict().then(messages => {
+            this.setState({messages});
+            this.predictor.terminate();
+        });
+    }
+
     render() {
-        if(false){
-            return <Layout />;
-        }else{
-            return <Loading />;        
-        }
+        return this.state.messages !== null ? 
+            <Layout messages={this.state.messages} /> : 
+            <Loading />;
     }
 }
